Fix how the CLI version is determined
We were trying to pass this in with an `npm_package_version` environment
variable before, but it wasn't reliable. This adds some utilities for
finding/loading the project's `package.json` file directly and pulls the
version out of that. These utilities will also be useful for other things like
determining the location of the scaffolding templates and finding local config
files.

Merges #18