Add a `lint` command and integrate the Sindri Manifest JSON Schema
This adds npm scripts to download the Sindri Manifest JSON Schema from the
backend. Currently it's stored at the root of the project. It can be updated by
running:

```bash
yarn download-sindri-manifest-schema
```

This also adds a `sindri lint` command which validates a project's
`sindri.json` file against the schema. It also checks for a `README.md` file
and logs it as a warning if it is not found. We can add more checks in the
future.

Merges #20