Add a `sindri deploy` command
This adds a basic `sindri deploy` command to upload and compile a circuit. The
generated OpenAPI client unfortunately didn't support passing in `FormData`
directly, so I made some small manual edits to the generated code... This is
unfortunate because it will be overwritten when we regenerate the client, I'll
loop back and come up with a better fix later.

Merges #22