Add support for Circom in `sindri init`
This adds support for Circom circuits in the `sindri init` with a basic
template scaffold checking that `X === Y`. There are also a few bug fixes that
slipped in. A notable thing is that I pulled in `circomlib` as a dependency in
the scaffolded project because I think people are likely to want it. Previously
we had been excluding files from `.gitignore` when deploying, but I think it
makes sense for `node_modules` to be in `.gitignore` but also be part of the
deploy here, so I changed that behavior to only check `.sindriignore` files.

Merges #24

LGTM given by: @katiemckeon