Disallow hyphens in Noir package names in `sindri init`
This updates the `sindri init` command logic around validating Noir package
names to disallow hyphens. Through trial and error, I found that it seems to be
fine with any mix of alphanumeric chacters and underscores, so I also relaxed
the validation around this. I think that the package name is never used in
code, unlike the Halo2 package names, and that's why it's OK for it to start
with a digit even though that wouldn't be a valid code identifier.

Merges #32

LGTM given by: @katiemckeon