Flesh out the README
This adds some content to the README file (which is used both on GitHub and
NPM). There's some minor polish like adding badges and a logo, but also a
stripped down version of the Quickstart guide. I wanted to at least provide a
minimal version of that while also continuously pointing readers to the real
README.

Closes #30
Merges #36

LGTM given by: @katiemckeon