Add IIFE build and browser/node testing framework
This configures an IIFE build that can be used to pull in the SDK library using
a script tag. It also sets up AVA as a test runner with three different
configurations:

- `test:dev` - Run the `test/lib/` tests from `src/lib/` without an explicit
build step.
- `test:node` - Run the `test/lib/` tests from `dist/lib/` requiring an
explicit build.
- `test:browser` - Run the `test/browser/` tests in Chromium using
`dist/lib/browser/sindri.iife.js` requiring an explicit build.

I don't think there's really a way to run the same test suite in both the
browser and node environments, so my plan is to do most of the testing in node
and have a basic set of sanity check tests that run in puppeteer to make sure
that the browser builds don't accidentally depend on node-only libraries or
lack any polyfills.

Merges #41