Simplify test setup
The test configuration was previously fairly complicated with different
`ava.config.*` and `tsconfig.*` files for testing against the source directory
and different build outputs. This collapses everything down to a single
configuration which runs all of the tests and only supports running them
against build outputs. This also allowed simplifying the npm scripts
significantly because fewer configurations are necessary. I added a `yarn
test:watch` command which automatically handles rebuilding and rerunning tests
after changes in development, so the development ergonomics here should be
reasonably good aside from needing to wait a couple of extra seconds for the
build. We can add back more complexity if/when we need, but I think it's better
to start out with a simpler configuration that requires less knowledge to use.

Merges #43