Use string unions instead of enums in the SDK
This updates the SDK to use string unions instead of enums for `ProofStatus`
and `CircuitStatus`. This allows us to have a single non-type export for the
SDK so that we can use the CJS/ESM interop capability of tsup in the build
process so that `require().default` isn't required.

```
> require('./dist/lib/index.js')
_class { pollingInterval: 1000 }
```

Merges #56

LGTM given by: @katiemckeon