Regenerate API with recent changes
This includes a handful of recent changes in the upstream API:

* `/user/me/` is a GET request instead of a POST. Minor shuffling of response
* fields (*e.g.* `num_proofs` added). New password change endpoint.

The `/user/me/` change is really the key one, this breaks compatibility with
previous releases of `sindri login`.

A `CommonCircuitInfoResponse` type is now included as a possible response type
in the schema, but this was only added as a workaround for a bug with
django-ninja. I opted to manually remove it from the generated API because
otherwise I would have had to change the public client types in a way that
doesn't reflect reality. Merges #65

LGTM given by: @KPreisner