Add node version testing matrix and specify compatible engines
This adds a testing matrix that covers node versions 18-21, and specifies the
minimum node engine version in `package.json` as v18.3.0. One blocker to
supporting earlier versions is that `buffer.File` was added in v18.3.0. I took
a stab at removing this, but ran into a lot of issues trying to get things to
work without it. v16 was EOLed last year though, so I don't think this is that
major of an issue. It will be easier to test against earlier versions now with
the testing matrix if we do decide we want to support them.

One notable thing is that the `--loader` vs `--import` argument depends on
versions in fairly complex ways. I think that it's roughly `--import` for
`>=18.19.0 < 19.0.0 || >=v20.6.0` and `--loader` for everything else, and I
have a commit in here with something like that working in the ava config, but
ultimately went with running ava with tsx like `tsx ./node_modules/.bin/ava`
because I'm more confident that this will work with any node version (despite
the fact that it's a bit hackier).

Closes #64
Merges #67