Improve Sindri Manifest lint output by discriminating circuit type
When there are Sindri Manifest schema issues, the linting output was previously
not as helpful as it could be because issues would be logged across all circuit
types. This happened because the validator wasn't able to accurately
discriminate the union of different subschemas. This PR manually discriminates
the circuit type and modifies the schema that we're validating against to be
specifically for the matching subschema. The code is a little verbose/hacky,
but the error messages are significantly more helpful so I think the tradeoff
is well worth it.

Merges #74