Integrate Circomspect into `sindri lint`
This integrates Circomspect into the `sindri lint` command such that its static
analysis errors and warnings are both logged and counted towards the total
counts for each category. It currently requires the Circomspect to be installed
locally on the machine, but the plan is to transparently pull down and run an
optimized docker image from docker-zkp.

Merges #76