Add `sindri exec|x` command with Circomspect support
This adds a new `exec` subcommand (or `x` for short) that allows transparently
pulling and executing ZK tools in a Docker container with the local project
mounted. [Circomspect](https://github.com/trailofbits/circomspect) is currently
the only available tool, but we'll expand this with more commands in the
[docker-zkp](https://github.com/sindri-labs/docker-zkp) repository. The `sindri
lint` command was also updated to use a locally installed version of
Circomspect if it's available, and to otherwise fall back to using Docker.

Merges #77