Fix `/tmp/` dir mounting on macos in docker
The `os.tmpdir()` call was returning a path under `/var/folders/` on macs, so
the Circomspect output was ending up in a different location than the CLI was
expecting it. This hardcodes the path to `/tmp/` which is effectively all we
support at the moment. Windows support outside of WSL will be more complicated
because it will require further path rewriting.

Merges #78