Improve logged Circomspect file paths when running in Docker
The `sindri lint` file paths logged for Circomspect output were rewriting the
paths relative to the project's root directory on the host system, but the
project is mounted at `/sindri/` inside of the container so this was resulting
in spurious `../../../` prefixes on the file paths. This adjusts how we rewrite
the paths differently based on whether or not the command is being run in
Docker.

Merges #79