Add ESM/CJS import tests
This adds some basic tests to make sure that the ESM and CJS builds are
working. It's easy to accidentally add a library that doesn't work with one of
the builds, or to do something that pushes the library under a `default`
property on the CJS build. I've added and reverted a few commits here that
demonstrate these issues and the tests catching them.

In particular, see these commits which introduce different errors:

* a182acf - Introduce an ESM-only dependency
* 019fd03 - Force the CJS build to export the library under `.default`
* 495f335 - Use a CJS only library as an ESM import

Closes #63

Merges #84

LGTM given by: @katiemckeon