Add `Sindri-Client` header to requests
This adds a `Sindri-Client` header to all SDK requests which specifies the
client type and version. We use separate identifiers for the SDK and CLI of the
form `sindri-js-sdk/v0.0.0` and `sindri-js-cli/v0.0.0` respectively. This will
allow us to keep track of which clients are people using the most. We were
previously pulling the version string from the `package.json` file in the
project, but this isn't available in the browser builds, so this PR updates it
to use a `VERSION` environment variable that is specified as part of the
deployment process.

Closes #86

Merges #87