Add a `sindri proof create` command
This adds a basic `sindri proof create` command for generating a proof for the
circuit. This currently just prints out the relevant subset of the API
response, we don't attempt to save the outputs to disk or integrate with local
tools. There's a `--verify` flag which optionally performs a remote
verification of a proof, and a `--tag` flag which allows proving against a
specific tag.

This update also required a regeneration of the internal API client and the
test fixtures, so there's a fair bit of noise in the diffs. The main files to
pay attention to are `src/lib/client.ts`, `src/cli/index.ts`, and
`src/cli/proof.ts`.

Closes #35
Merges #90

LGTM given by: @katiemckeon