Add a CI/CD GitHub Actions Workflow to the circuit templates
This adds a new `templates/common/.github/workflows/ci.yaml` template to the
project scaffolds which includes a basic CI/CD workflow. On pull requests, this
will lint the circuit, compile it remotely, and then create a proof remotely
and verify it. When pull requests are merged into `main` or `master`, it will
run `sindri deploy` to deploy the `latest` version of the circuit. Git tags
that start with `v` also trigger circuit deploys with the same tag as the git
tag. The one non-ideal part is that there isn't an obvious circuit tag to use
for a temporary compiled circuit when doing tag releases, so the prove step is
skipped in this case.

Merges #91