Add a `sindri clone` CLI command
This adds a `sindri clone` command which can be used to download either your
own circuit or somebody else's public circuit. You can either specify an exact
circuit UUID or DockerHub inspired identifiers like `<circuit-name>`,
`<circuit-name>:<tag>`, `<team>/<circuit-name>`, or
`<team>/<circuit-name>:<tag>`.

I haven't added it to the SDK yet, we'll need to think about a good interface
to simultaneously support browser and node there. The auto-generated internal
API client unfortunately required a number of modifications in order to support
streaming responses which are necessary for very large circuit packages. These
are starting to build up, and we need to think about ways to generalize these
diff patches as part of the generation process. The
`openapi-typescript-codegen` codebase is unfortunately really complicated, so
it's not practical to fork or patch it directly.

Merges #95

LGTM given by: @jason-sindri