Remove `getAllProofs()` and deprecated `include_*` API parameters
This regenerates the internal API client based on recent upstream changes. This
includes making `/api/v1/proof/list` an internal endpoint, adding `public` to
circuit info responses, and various documentation improvements. Because listing
all proofs is now an internal-only method, I removed
`SindriClient.getAllProofs()`. We were using this in the `getProof()` tests, so
these also had to be updated not to use it.

Closes #96
Merges #97