Bump Puppeteer to v22.6.0 and Chromium to v123
This bumps Puppeteer for the browser tests from v22.2.0 to v22.6.0 and Chromium
from v122 to v123. The docker development image on ARM requires pulling in an
ARM Chromium build from a third-party source (Microsoft Playwright), and this
was updated to
[r1107](https://github.com/microsoft/playwright/commit/f8e441a52182043956c7277e48467e2d9230d796)
which provides Chromium v123.0.6312.22. This differs slightly from the
v123.0.6312.58 version bundled with Puppeteer, but it's not consequential.
Playwright already has a v124 build in
[r1110](https://github.com/microsoft/playwright/commit/352cb7fa08deaad87a594811f24306e509f83061),
but Puppeteer doesn't have an explicitly compatible v124 release yet. In
reality, minor version mismatches don't really matter for what we're doing
here, but I wanted to try to keep the versions as close as possible for now.

For future reference:

- https://pptr.dev/chromium-support - Provides Puppeteer/Chromium version
compatibility.
- https://github.com/microsoft/playwright/commits/main/ - Dig through here for
updates to the Chromium version, then update the release number in the ARM
Chromium download URL in the Dockerfile.
-

Merges #98