Add node v22 to the test matrix on CI
Node v22 was released yesterday and is the new "current" release slated to
become the next LTS release. This adds it to the CI test matrix.

A couple things came up while doing this:

One is that ava doesn't yet specify node v22 as a supported engine, but it also
doesn't specify node v19 or v21, so we already had this issue but weren't
failing CI because of it. I tried setting up strict engine checks only for
production dependencies (see 07d30a5), but it seems that npm checks the dev
dependencies even when specifying `--omit=dev`. We should really fail CI if a
production dependency doesn't support the given engine, but this is somewhat
unrelated to the v22 tests and I'd like to do it in a followup when I have more
time to think of a good solution.

Another issue that came up when installing only the production dependencies is
that we use patch-package to patch http-mitm-proxy for tests, but this is only
a development dependency and patch-package fails when it is absent. This is
fine in the production package because we don't include the patches directory,
but it's not super obvious that patches *aren't* part of the package and it's
easy to accidentally add a new patch for a production dependency that will run
fine on CI but be missing from the deployed package. We also don't need
patch-package as a production dependency with the current setup if we can
conditionally run it in the postinstall script. I'd similarly like to follow up
on this and figure out how to improve the situation.

Merges #102

LGTM given by: @jason-sindri