Fix multiple deploy tag handling in the CLI with an explicit project directory
The `--tag` option on `sindri deploy` was defined as a variadic option
`<tag...>` which allowed specifying multiple tags. This created a problem when
the optional positional argument for the project directory was also specified
because it would be interpreted as an additional tag. For example, `sindri
deploy -t v1.0.0 my-circuit` would treat both `v1.0.0` and `my-circuit` as tags
while using the default `.` as the project directory.

This PR fixes the handling so that you need to repeat the tag argument for each
tag and the project directory is no longer consumed as a tag. So you can now do
`sindri deploy -t v1.0.0 -t latest my-circuit` and have `v1.0.0` and `latest`
treated as tags while `my-circuit` is the project directory.

Merges #108

LGTM given by: @jason-sindri