Set default circomspect library paths when versions allow
Circomspect added support for a `-L`/`--library` search path flag in analogy to
Circom's `-l` argument (as of trailofbits/circomspect#21) that we use on the
backend with defaults of the project root and the `node_modules` subdirectory.
This PR updates the `sindri lint` command to check whether the current
Circomspect version supports the argument, and then to specify the same search
paths if the argument is supported. We'll probably end up making the search
paths configurable, but this makes the backend consistent with `sindri lint` in
the meantime.

For testing it manually, you can run

```bash
docker compose up -d
docker compose exec sindri-js bash
```

on the latest branch, change into a directory with one of our example Circom circuits that has been failing linting, and then run:

```bash
sindri lint
```

to make sure that things work as expected.

Merges #116

LGTM given by: @KPreisner