Add request retries to `SindriClient`
This adds a second argument to the `SindriClient()` constructor which can
include a `retryOptions` field which controls the newly added retry behavior.
The options are passed through to the `retry` package and are documented
[here](https://www.npmjs.com/package/retry#retrytimeoutsoptions). By default,
we'll retry 4 times with exponential backoff starting at a 1000 ms timeout. The
logic to determine whether to retry is hard-coded, we retry for status codes of
502-504 or any sort of connection errors.

Note that we use a fork of `p-retry` here which supports both ESM and CJS because we provide builds for each in our package (see: sindresorhus/p-retry/issues/76).

Closes #118

Merges #120

LGTM given by: @katiemckeon