Add support for attaching metadata to proofs and circuit builds
This adds support for specifying arbitrary key/value string pairs as metadata
attached to a circuit or proof. This can be used for a variety of purposes like
tracking which individual user on a team submitted a proof or storing database
IDs from an external system. In addition to being able to explicitly specify
metadata in `SindriClient` or the CLI, there's also support for a `SINDRI_META`
environment variable which can specify default metadata that will be included
in all requests. This variable can either be formatted as a JSON object (*e.g.*
`SINDRI_META='{"key": "value"}'`) or a colon delimited sequence of assignments
(*e.g.* `SINDRI_MANIFEST=key1=value1:key2=value2`). The latter format is
designed to be a little more shell friendly, so you can do things like add
`export SINDRI_META="user=$(whoami):host=$(hostname)` to your `.bashrc` or
`SINDRI_META="${SINDRI_META}:experiment=A" sindri deploy"`. In that format,
literal colons can be escaped as `::`.

When using `SindriClient` directly or the CLI, validation of meta keys in
values is performed: 1) both must be strings, 2) keys must be 1-64 characters,
3) keys must start with an alphabet character and include only alphanumeric
characters, hyphens, and underscores, and 4) values must be 0-4096 characters.
Using the client directly will always raise an exception if either
`SINDRI_META` or the data passed in `meta` is invalid. The CLI will raise an
exception if one of the explicity `--meta` arguments is invalid, but will log a
warning and ignore any invalid entries in `SINDRI_META`. The reason for the
difference in behavior here is that `SINDRI_META` is evaluated when
`src/cli/(proof|deploy).ts` are imported so that it can be shown as the default
value in the help, so an exception there would prevent the help from being
printed out. It's a little weird that they're different, but I think it's OK.

To use it with the CLI, you can run commands like:

```bash
# These should work:
sindri deploy -m key1=value1 --meta key2=value2
SINDRI_META="user=$(whoami)" sindri deploy --meta key=value
SINDRI_META="value-with-colons=a::b::c" sindri proof create

# These should log warnings or fail:
SINDRI_META=not-an-assignment sindri deploy
SINDRI_META='invalid!key=value' sindri proof create
sindri deploy -m not-an-assignment
```

Merges #124

LGTM given by: @jason-sindri