Bump Puppeteer to v23.2.1 and Chromium to v128
This bumps Puppeteer for the browser tests from v22.6.0 to v23.2.1 and Chromium
from v123 to v128. The docker development image on ARM requires pulling in an
ARM Chromium build from a third-party source (Microsoft Playwright), and this
was updated to
[r1129](https://github.com/microsoft/playwright/commit/6c6f10b67807eb8b36b0dfcf2b278735269e82ce)
which provides Chromium v128.0.6613.18. This differs slightly from the
v128.0.6613.86 version bundled with Puppeteer, but it's not consequential.
Minor version mismatches don't really matter for what we're doing here, but
it's good to keep them as close as possible.

For future reference:

- https://pptr.dev/chromium-support - Provides Puppeteer/Chromium version
compatibility.
- https://github.com/microsoft/playwright/commits/main/ - Dig through here for
updates to the Chromium version, then update the release number in the ARM
Chromium download URL in the Dockerfile.

Connects #98

Merges #145

LGTM given by: @jason-sindri