Fix local dev container installation of Puppeteer with amd64
We already jump through some hoops to install Chromium on arm using the PlayWright builds. On amd64, Puppeteer is supposed to install Chrome by default when you install the package, but this doesn't happen when we run `npm install` in the container run command for some reason. This updates the `Dockerfile` to manually install the same version of Puppeteer when we're not on arm to force the browser to be cached in `~/.cache/puppeteer/`. It's odd that this is required, but it's actually nice to have the browser included in the image regardless of architecture for consistency.

Merges #149

LGTM given by: @KPreisner