Adding Jolt & Sp1 examples to `sindri init` templates (#187)
This adds example guest code for jolt and Sp1 to our circuit templates.
While the Sp1 template is minimal as there are few configuration options
that affect the high level structure of guest code, the Jolt template is
quite dependent on the value of `stdEnabled` value. If you set that to
true, the template makes use of the rust standard library string
functions. Jolt code must follow [certain
patterns](https://jolt.a16zcrypto.com/usage/stdlib.html) depending on
whether that is enabled. For this reason, the template syntax
surrounding `stdEnabled` in `templates/jolt` should be examined/tested
carefully.

---------

Co-authored-by: Evan Sangaline <esangaline@gmail.com>