Add a `sindri project delete` command
This adds a new `sindri project delete` subcommand which can be used to delete
either a specific project or all projects recursively. The main motivation is
that we record test fixtures against the production API and they've been
accumulating a lot of cruft over time, so now we can call `sindri project
delete --all` on CI to get a clean slate before recording new test fixtures.

There's starting to be some code duplication in the CLI commands around things
like finding and parsing a project's `sindri.json` file. We should aim to
refactor and clean some of this up, but I think it's manageable for now.

Merges #208

LGTM given by: @katiemckeon