Properly surface non-serializable errors in tests
AVA runs tests in separate processes, and it uses IPC to communicate them back
to the main process for logging and reporting. If an error isn't serializable,
then the IPC fails and we end up not being able to see the actual error or even
what test failed which is as delightful as it sounds. The main culprit of this
has been Axios when there's a Nock error.

This PR improves the situation by adding an Axios response interceptor and
sanitizing any errors so that they're serializable. The heavy lifting on the
sanitization is done by the serialize-error package which I've added as a dev
dependency. It cleans up issues like functions, buffers, and circular
references which are the common culprits of non-serializability.

If you want to test this yourself, you can remove fixtures with `rm
test/fixtures/sdk.test.ts.json` and then run `npm run test` to run the tests.
You'll get the opaque error on `main` and the real error on this branch.

Merges #212

LGTM given by: @katiemckeon