Make tests stateless and request boundary handling more robust with fixtures
Following #208, we started recursively deleting all projects before
re-recording test fixtures. A couple of the tests were relying on some circuits
to have been compiled previously, and this created a race condition where these
tests would fail if they ran before the others. This PR updates those tests to
explicitly create their own circuits and look for those in the retrieved
circuit list.

There was also a race condition that was introduced in #175 with the avatar
upload test, but it never came up until we slimmed down the fixtures and it
changed timings and it seems to exacerbated by the other fix in the PR. The
issue here was that we were juggling the `Content-Type` header like we do with
circuit uploads so that the multipart form boundary would match, but this whole
juggling procedure was imperfect and had a TODO in place to replace it. The new
version stops trying to override the built in randomized form boundaries and
relies solely on the boundary regex. We already had pretty robust and
complicated handling of the dynamic boundaries in the request bodies, so I
mostly needed to update the regex to support the current Axios boundary format
and clean out the cruft.

Merges #211

LGTM given by: @katiemckeon