Fix team selection in `sindri login` command
The API key generation endpoint schema changed recently to explicitly include a
`Sindri-Team-Id` header. The actual endpoint behavior didn't change, but the
OpenAPI schema change caused an explicit `sindriTeamId` argument to be added to
`apikeyGenerate()` in the client, and the default value of `null` was
overriding the header that we were setting explicitly in the `sindri login`
command. This caused all API key generation to default to a user's personal
team regardless of the team they selected. This PR updates `sindri login` to
use the new `sindriTeamId` argument.

Merges #218

LGTM given by: @jason-sindri