Disable the GPU in the browser tests
We've been getting this error intermittently on some of our CI runners:

```
ProtocolError {
message: 'Runtime.callFunctionOn timed out. Increase the \'protocolTimeout\'
setting in launch/connect calls for a higher timeout if needed.',
name: 'ProtocolError',
}
```

Some others have reported getting around this by disabling the GPU (see:
https://github.com/puppeteer/puppeteer/issues/12637#issuecomment-2264825922).
This PR gives that a shot and we'll see if the issues disappear.


Merges #229